fix: 좋아요 여러 번 클릭했을 때 수정

+++ frontend/src/components/Card/CardRecruit.jsx
@@ -1,16 +1,15 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import styles from './Card.module.css';
 import Button from '../Button/Button';
 import heartRed from './images/heart_red.svg';
 import heartGray from './images/heart_gray.svg';
 import { v4 as uuidv4 } from 'uuid';
-import { Context } from '../../context';
 import { apiClient } from '../../api/api';
 
-const CardRecruit = ({ post }) => {
+const CardRecruit = ({ post, user }) => {
   const { author, tags, title, content, likeMembers, image } = post;
-  const [state, dispatch] = useContext(Context);
   const [profileImgURL, setProfileImgURL] = useState(null);
+  const [mount, setMount] = useState(true);
   const [like, setLike] = useState(post.like);
   const [likes, setLikes] = useState(likeMembers.length);
   const IMG_REGISTER_URL = `http://localhost:4000/api/auth/${author?._id}/profile-image`;
@@ -25,22 +24,31 @@ const CardRecruit = ({ post }) => {
   const likeHandler = async (e) => {
     e.preventDefault();
     if (like) {
-      const response = await apiClient.delete(`/api/posts/${post._id}/likes`);
-      setLikes(response.data.length);
+      setLikes((prev) => prev - 1);
     } else {
-      const response = await apiClient.post(`/api/posts/${post._id}/likes`);
-      setLikes(response.data.length);
+      setLikes((prev) => prev + 1);
     }
 
     setLike((prev) => !prev);
   };
 
   useEffect(() => {
+    if (mount) {
+      setMount(false);
+      return;
+    }
+
+    if (like) {
+      apiClient.post(`/api/posts/${post._id}/likes`);
+    } else {
+      apiClient.delete(`/api/posts/${post._id}/likes`);
+    }
+  }, [like]);
+
+  useEffect(() => {
     if (author?._id) {
       getProfileImage();
     }
-
-    setLike(likeMembers.indexOf(state.user?._id) !== -1);
 
     return () => setProfileImgURL(profileImgURL);
   }, []);
@@ -90,7 +98,7 @@ const CardRecruit = ({ post }) => {
               text={likes}
               ftsize='1.6rem'
               onClick={
-                state.user === null
+                user === null
                   ? (e) => {
                       e.preventDefault();
                       alert('회원만 사용할 수 있는 기능입니다.');
